Fix order in which server binary is found in vscode extension
Previously we would always prefer a `zeek-language-server` binary found
in PATH over a configured one; with this patch we now fix the order so a
configured one takes precedence.

We also remove some dead code looking for a server binary with a
platform-specific name. Such a binary is not used anymore since some
time.