Do not attempt to start server on unsupported platforms
This extension now contains more than just support for running the
server, so it makes sense to use it on other platforms as well. On such
platforms we would still have emitted an error message about being
unable to download the server, and also would have performed a check for
`zeek-format` which is useless without the server

With this patch we now skip the server initialization on such platforms
and emit a simple information message.