Perform file list update in `did_change_watched_files` under wider lock
For vim-lsc we seem to run into caseswhere the initial
`did_change_watched_files` seems to race with a `did_open` from the
client on updating the list of files; it seems that for such setups
ultimately only the file from `did_open` seems to make it into the final
list.

With this patch we move the updating of the file list from
`did_change_watched_files` under a wider lock so the full computation
and update of the list happens under a write lock. This should properly
sequence concurrent updates.

Closes #52.