Tweak resolver for identifiers in same module, but other file
We previously assumed that an identifier either refers to a declaration
in the same file, or is fully qualified if it does not reference a
declaration in the `GLOBAL` module. This does not hold in practice since
Zeek modules can be extended in any later file.

This patch adds an additional fallback resolve step if a name cannot be
resolved where we turn an unqualified into a fully-qualified identifier.
We perform this after the normal lookup to not interfere with the lookup
for e.g., local identifiers which also have no module.

Closes #57.