Fix response for formatting request
We previously would incorrectly emit both a deletion and an insertion
of the formatted code. This was incorrect in that their ranges by
construction overlapped (not allowed per spec).

We now emit a replacement edit where the existing source range gets new
text. This also improves or fixes formatting in editors, e.g, coc.vim
would previously ignore our edit, and vscode would accept it, but have
the cursor jump.