Improve rendering when completing impls for function-likes
When completing implementation bodies for events/functions/hooks we
would previously return the full signature in the completion label. This
can be a very long string for e.g., some events.

With this path instead instead only show the name in the label, and the
signature in additional details. Some clients show this in a muted color
while others do not expose this information at all. We still also return
the full signature in the doc string.