Prefer to resolve IDs to declarations instead of definitions
For constructs which can have separate declarations and definitions
(events, functions, hooks) when resolving IDs we previously would always
return the last instance. Most of the time this meant a definition and
not the declaration which at least in the case of Zeek base scripts
has the documentation.

We now prefer to return the definition instead. This fixes gotodef for
e.g., event handler implementations. We also now show to docstring from
the declaration which most of the time is more useful (user scripts
contain many implementations of events declared and documented
elsewhere).