Use a per-file database to store source code
We previously would track the list of known files independently of
sources of files. This made it potentially possible that we would call
`source` for an unknown file (which would then panic).

Instead of adding manual checking everywhere where `source` is called,
this patch instead introduce separate a database for each file which can
be accessed via `files`. This makes it impossible to get sources for
unknown files.

We also clean up the `Default` implementation for our global `Database`
so that it automatically sets storage for its cached containers whenever
the a default-constructed.