Make request handling in the server sequential
This is the brute force method of removing deadlocks in request
handling. Previously we sometimes could have mutating accesses to the
database block since snapshots were still in flight; if the snapshots
lived on the same thread as the mutating access this could deadlock. By
making the server handle at most one request at a time this error
scenario is eliminated.