Use snapshots in smaller scopes
If any database snapshots are alive when a `set` operation is performed
on the database, the operation will block. This has been a constant
issue with a server with `concurrency_level>1` and e.g., an initial
index racing with incoming files changes from editing.

By using snapshots in smaller scopes we dramatically reduce the
likelihood of this happening, and can make the server more concurrent in
a later commit.