Fix deadlock
We would hold onto a read lock on the state for the full duration of
`did_change_watched_files`. This function only finishes once all
snapshots held by parallelized tasks are dropped which could themself
block if we tried to acquire a write lock on the state elsewhere, e.g.,
from `did_change`.

With this patch we acquire the lock in `did_change_watched_files` for a
much shorter time. This seems to remove deadlocks we previously saw for
a concurrent server (reproducible by editing while the initial index was
in flight), so we make the server concurrent again.