Simplify query
This query had a lot of states which can cause constructing it to be
_very_ expensive (in my local profiling about 30% of total startup
time). I believe this is due to
https://github.com/tree-sitter/tree-sitter/issues/973 which has been
sitting for a while.

As a workaround strip the query down to what we really need. As we
remove more and more stuff the danger is that we might end up matching
unrelated code, but what is here now seems to be sufficient.